Refactor Markdown and Mermaid Rendering Components
- Removed the existing MarkdownRenderer and MermaidRenderer components along with their associated styles.
- Introduced a new Renderer module that consolidates functionality for detecting Markdown, Mermaid diagrams, and code content.
- Implemented a new MarkdownRenderer component for simple Markdown-to-HTML conversion.
- Added utility hooks for managing stage transitions and auto-scrolling behavior in the chat interface.
- Enhanced message handling to support Markdown rendering based on artifact stages and sender type.
- Created helper functions for determining the current action description and greeting text based on the artifact stage.
- Established a new structure for managing stage transitions and user interactions in the chat application.
