+++ src/components/ChatContainer/ChatContainer.css
@@ -11,6 +11,8 @@
     margin-bottom: 0.5rem;
     padding: 0.6rem 1rem;
     border-radius: 1rem;
+    word-wrap: break-word;
+    overflow-wrap: break-word;
 }
 
 .chat-bubble.user {
@@ -94,6 +96,7 @@
 .chat-bubble.bot .markdown-p {
     margin: 8px 0;
     font-size: 1em;
+    line-height: 1.4;
 }
 
 .chat-bubble.bot .markdown-ul,
@@ -104,6 +107,7 @@
 
 .chat-bubble.bot .markdown-li {
     margin: 4px 0;
+    line-height: 1.3;
 }
 
 .chat-bubble.bot .markdown-code-block {
@@ -113,6 +117,9 @@
     padding: 12px;
     margin: 8px 0;
     font-size: 0.9em;
+    font-family: 'Courier New', Courier, monospace;
+    overflow-x: auto;
+    white-space: pre-wrap;
 }
 
 .chat-bubble.bot .markdown-inline-code {
@@ -121,6 +128,7 @@
     border-radius: 3px;
     padding: 1px 4px;
     font-size: 0.9em;
+    font-family: 'Courier New', Courier, monospace;
 }
 
 .chat-bubble.bot .markdown-blockquote {
@@ -129,4 +137,56 @@
     margin: 8px 0;
     padding: 8px 12px;
     border-radius: 0 4px 4px 0;
+    font-style: italic;
+}
+
+/* Enhanced styles for better code and diagram rendering */
+.chat-bubble.bot .markdown-mermaid {
+    background-color: #ffffff;
+    border: 1px solid #dee2e6;
+    border-radius: 6px;
+    padding: 12px;
+    margin: 8px 0;
+    text-align: center;
+    overflow-x: auto;
+}
+
+.chat-bubble.bot .markdown-table {
+    border-collapse: collapse;
+    width: 100%;
+    margin: 8px 0;
+    font-size: 0.9em;
+}
+
+.chat-bubble.bot .markdown-table th,
+.chat-bubble.bot .markdown-table td {
+    border: 1px solid #dee2e6;
+    padding: 6px 8px;
+    text-align: left;
+}
+
+.chat-bubble.bot .markdown-table th {
+    background-color: #e9ecef;
+    font-weight: bold;
+}
+
+/* Mobile responsiveness */
+@media (max-width: 768px) {
+    .chat-bubble {
+        max-width: 85%;
+        padding: 0.5rem 0.8rem;
+    }
+    
+    .chat-container {
+        padding: 0.5rem;
+    }
+    
+    .chat-actions {
+        flex-direction: column;
+        gap: 0.5rem;
+    }
+    
+    .chat-actions button {
+        width: 100%;
+    }
 }
