--- src/components/MarkdownRenderer/MarkdownRenderer.css
@@ -1,147 +1,0 @@
-/* Base markdown renderer styles */
-.markdown-renderer {
-    line-height: 1.6;
-    word-wrap: break-word;
-}
-
-/* Empty line spacing */
-.markdown-empty-line {
-    height: 0.5em;
-}
-
-/* Headers */
-.markdown-h1 {
-    font-size: 1.8em;
-    font-weight: bold;
-    margin: 16px 0 8px 0;
-    border-bottom: 2px solid #48dbfb;
-    padding-bottom: 4px;
-    color: #1e3a5f;
-}
-
-.markdown-h2 {
-    font-size: 1.5em;
-    font-weight: bold;
-    margin: 14px 0 7px 0;
-    border-bottom: 1px solid #48dbfb;
-    padding-bottom: 2px;
-    color: #1e3a5f;
-}
-
-.markdown-h3 {
-    font-size: 1.2em;
-    font-weight: bold;
-    margin: 12px 0 6px 0;
-    color: #1e3a5f;
-}
-
-/* Paragraphs */
-.markdown-p {
-    margin: 8px 0;
-    line-height: 1.6;
-}
-
-/* Lists */
-.markdown-ul,
-.markdown-ol {
-    margin: 8px 0;
-    padding-left: 20px;
-}
-
-.markdown-li {
-    margin: 4px 0;
-    line-height: 1.5;
-}
-
-/* Code blocks */
-.markdown-code-block {
-    background-color: #f8f9fa;
-    border: 1px solid #e9ecef;
-    border-radius: 6px;
-    padding: 12px;
-    margin: 12px 0;
-    font-family: 'Courier New', Courier, monospace;
-    font-size: 0.9em;
-    overflow-x: auto;
-    white-space: pre;
-}
-
-.markdown-code-block code {
-    background: none;
-    border: none;
-    padding: 0;
-    font-size: inherit;
-    font-family: inherit;
-}
-
-/* Inline code */
-.markdown-inline-code {
-    background-color: #f8f9fa;
-    border: 1px solid #e9ecef;
-    border-radius: 3px;
-    padding: 2px 4px;
-    font-family: 'Courier New', Courier, monospace;
-    font-size: 0.85em;
-}
-
-/* Text formatting */
-.markdown-bold {
-    font-weight: bold;
-}
-
-.markdown-italic {
-    font-style: italic;
-}
-
-.markdown-strikethrough {
-    text-decoration: line-through;
-}
-
-/* Links */
-.markdown-link {
-    color: #007bff;
-    text-decoration: underline;
-    cursor: pointer;
-}
-
-.markdown-link:hover {
-    color: #0056b3;
-    text-decoration: none;
-}
-
-/* Blockquotes */
-.markdown-blockquote {
-    background-color: #f8f9fa;
-    border-left: 4px solid #48dbfb;
-    margin: 12px 0;
-    padding: 8px 12px;
-    border-radius: 0 4px 4px 0;
-    font-style: italic;
-}
-
-/* Horizontal rules */
-.markdown-hr {
-    border: none;
-    border-top: 2px solid #e9ecef;
-    margin: 20px 0;
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-    .markdown-h1 {
-        font-size: 1.5em;
-    }
-    
-    .markdown-h2 {
-        font-size: 1.3em;
-    }
-    
-    .markdown-h3 {
-        font-size: 1.1em;
-    }
-    
-    .markdown-code-block {
-        padding: 8px;
-        font-size: 0.8em;
-    }
-}
