+++ src/hooks/useAutoScroll.js
@@ -1,0 +1,20 @@
+import { useState, useEffect, useRef } from 'react';
+
+export function useAutoScroll(messages) {
+    const mainContainerRef = useRef(null);
+    const prevMessagesLength = useRef(0);
+
+    useEffect(() => {
+        if (messages.length > prevMessagesLength.current) {
+            if (mainContainerRef.current) {
+                mainContainerRef.current.scrollTo({
+                    top: mainContainerRef.current.scrollHeight,
+                    behavior: 'smooth'
+                });
+            }
+        }
+        prevMessagesLength.current = messages.length;
+    }, [messages]);
+
+    return mainContainerRef;
+}
